requirements.txt
numpy==1.26.4
faiss-cpu==1.8.0.post1
onnxruntime==1.18.0
insightface==0.7.3
opencv-python==4.10.0.84
tqdm==4.66.4
scikit-learn==1.4.2
